--- a/.changeset/fluffy-bobcats-tell.md
+++ b/.changeset/fluffy-bobcats-tell.md
@@ -1,0 +1,5 @@
+---
+'@backstage/eslint-plugin': patch
+---
+
+Fix custom rules package scanning performance.