@GabDug 2025-05-27
> Cool! In general - don't be afraid of using the yaml and globby packages. The robustness is a good tradeoff for the bytes otherwise saved.

@freben To ease maintenance and evolutions, I've reverted the changes with the custom implementation . Even though it could bring more performance,`manypkg` already supports multiple workspace managers with acceptable scanning performance.