@freben 2025-06-03
The test setup has received some fixes lately. A rebase on latest master should help things go green.