@drodil 2025-06-10
Also a bit related issue: https://github.com/backstage/backstage/issues/30190