@GabDug 2025-06-10
@freben I've just rested: with this MR (cache invalidation only when fixes are applied), I go from ~17s to ~9s.

I switched to a more powerful device, so I can't replicate the original benchmarks from the description.

I've just rebased and removed extra changes to keep only the change in cache invalidation.